rename page components
