+++ services/ui/react/src/App.tsx
@@ -1,22 +1,22 @@
 import { Routes, Route } from 'react-router';
 import { AuthProvider } from './AuthProvider/AuthProvider.tsx';
 import { Home } from './Home/Home';
-import { Welcome } from './Welcome/Welcome';
-import { Login } from './Login/Login';
+import { WelcomePage } from './WelcomePage/WelcomePage';
+import { LoginPage } from './LoginPage/LoginPage.tsx';
 import { NotFound } from './NotFound/NotFound';
-import { Users } from './Users/Users';
-import { Training } from './Training/Training';
+import { UsersPage } from './UsersPage/UsersPage';
+import { TrainingPage } from './TrainingPage/TrainingPage.tsx';
 import { SwaggerPage } from './Swagger/SwaggerPage.tsx';
 
 export const App = () => {
   return (
     <AuthProvider>
       <Routes>
-        <Route path="/login" element={<Login />} />
+        <Route path="/login" element={<LoginPage />} />
         <Route element={<Home />}>
-          <Route index element={<Welcome />} />
-          <Route path="training" element={<Training />} />
-          <Route path="users" element={<Users />} />
+          <Route index element={<WelcomePage />} />
+          <Route path="training" element={<TrainingPage />} />
+          <Route path="users" element={<UsersPage />} />
           <Route path="swagger" element={<SwaggerPage />} />
         </Route>
         <Route path="*" element={<NotFound />} />
